create appeal manager
