+++ Freelance_WbSellerFeedbackPareser/WbAppealsManager/WbAppealReader.cs
@@ -4,15 +4,15 @@ namespace Freelance_WbSellerFeedbackPareser.WbAppealsManager
 {
     internal class WbAppealReader
     {
-        private readonly IRequestSender _requestSender;
+        private readonly WbAppealManager _wbAppealManager;
         public WbAppealReader(IRequestSender requestSender)
         {
-            _requestSender = requestSender;
+            _wbAppealManager =new WbAppealManager(requestSender);
         }
 
         public string GetProductId(int appealId)
         {
-            string content = ReadAppeal(appealId);
+            string content = _wbAppealManager.ReadAppeal(appealId);
             List<AppealAttributeModel> appealAttributes = ConvertToAttributes(content);
             AppealAttributeModel? productData = appealAttributes
                 .FirstOrDefault(x => x.Name == "Номенклатура" || x.Name == "Артикул  товара");
@@ -24,14 +24,6 @@ namespace Freelance_WbSellerFeedbackPareser.WbAppealsManager
             {
                 return productData.Value;
             }
-        }
-
-        private string ReadAppeal(int appealId)
-        {
-            string url = $"https://seller.wildberries.ru/ns/suppliers-proxy/callcenter/suppliers-appeals-api/v1/suppliers/users/appeals/{appealId}";
-            var content = _requestSender.SendRequestAsync(HttpMethod.Get, url).Result;
-            Thread.Sleep(400);
-            return content;
         }
 
         private List<AppealAttributeModel> ConvertToAttributes(string content)
